Fixed button bar hiding permissions screen
The button bar uses windowbackground, which in theory can have
an intrinsic size which affects the size of the bar. Using the
color resource used by default for the windowbackground instead
removes such concerns.

Test: Manual, download .apk file (that requests permissions),
to /sdcard, and do installation from storage settings menu.
Observe that it is possible to install application.

Bug: 77617753

Change-Id: I527b918f699695801e7a35eca262b721c9f6356e